backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
supabase==2.8.1
alpaca-py==0.12.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
PyJWT==2.8.0